+++ notify/webhook/webhook.go
@@ -1,0 +1,111 @@
+// Copyright 2019 Prometheus Team
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package webhook
+
+import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"fmt"
+	"net/http"
+
+	"github.com/go-kit/kit/log"
+	"github.com/go-kit/kit/log/level"
+	commoncfg "github.com/prometheus/common/config"
+	"github.com/prometheus/common/version"
+
+	"github.com/prometheus/alertmanager/config"
+	"github.com/prometheus/alertmanager/notify"
+	"github.com/prometheus/alertmanager/template"
+	"github.com/prometheus/alertmanager/types"
+)
+
+var userAgentHeader = fmt.Sprintf("Alertmanager/%s", version.Version)
+
+// Notifier implements a Notifier for generic webhooks.
+type Notifier struct {
+	conf   *config.WebhookConfig
+	tmpl   *template.Template
+	logger log.Logger
+	client *http.Client
+}
+
+// New returns a new Webhook.
+func New(conf *config.WebhookConfig, t *template.Template, l log.Logger) (*Notifier, error) {
+	client, err := commoncfg.NewClientFromConfig(*conf.HTTPConfig, "webhook")
+	if err != nil {
+		return nil, err
+	}
+	return &Notifier{
+		conf:   conf,
+		tmpl:   t,
+		logger: l,
+		client: client,
+	}, nil
+}
+
+// Message defines the JSON object send to webhook endpoints.
+type Message struct {
+	*template.Data
+
+	// The protocol version.
+	Version  string `json:"version"`
+	GroupKey string `json:"groupKey"`
+}
+
+// Notify implements the Notifier interface.
+func (n *Notifier) Notify(ctx context.Context, alerts ...*types.Alert) (bool, error) {
+	data := notify.GetTemplateData(ctx, n.tmpl, alerts, n.logger)
+
+	groupKey, err := notify.ExtractGroupKey(ctx)
+	if err != nil {
+		level.Error(n.logger).Log("err", err)
+	}
+
+	msg := &Message{
+		Version:  "4",
+		Data:     data,
+		GroupKey: groupKey.String(),
+	}
+
+	var buf bytes.Buffer
+	if err := json.NewEncoder(&buf).Encode(msg); err != nil {
+		return false, err
+	}
+
+	req, err := http.NewRequest("POST", n.conf.URL.String(), &buf)
+	if err != nil {
+		return true, err
+	}
+	req.Header.Set("Content-Type", "application/json")
+	req.Header.Set("User-Agent", userAgentHeader)
+
+	resp, err := n.client.Do(req.WithContext(ctx))
+	if err != nil {
+		return true, err
+	}
+	notify.Drain(resp)
+
+	return n.retry(resp.StatusCode)
+}
+
+func (n *Notifier) retry(statusCode int) (bool, error) {
+	// Webhooks are assumed to respond with 2xx response codes on a successful
+	// request and 5xx response codes are assumed to be recoverable.
+	if statusCode/100 != 2 {
+		return (statusCode/100 == 5), fmt.Errorf("unexpected status code %v from %s", statusCode, n.conf.URL)
+	}
+
+	return false, nil
+}
